requirements.txt
aiohttp==3.9.3
asyncio==3.4.3
bs4==0.0.2
environs==11.0.0
python-telegram==0.18.0